+++ crypto.h
@@ -26,11 +26,11 @@ freely, subject to the following restrictions:
 
 int select_crypto_if(const int ifno);
 const char *crypto_init(const int ifno);
-unsigned long encrypt(const void *key, const int keylen, const void *pt, void *ct, const int tlen);
-unsigned long decrypt(const void *key, const int keylen, const void *ct, void *pt, const int tlen);
-unsigned long hash(const void *pt, const int tlen, void *tag, int *taglen);
-unsigned long hmac(const void *key, const int keylen, const void *pt, const int tlen,
-			void *tag, int *taglen);
+unsigned long encrypt(const void *key, const size_t keylen, const void *pt, void *ct, const size_t tlen);
+unsigned long decrypt(const void *key, const size_t keylen, const void *ct, void *pt, const size_t tlen);
+unsigned long hash(const void *pt, const size_t tlen, void *tag, size_t *taglen);
+unsigned long hmac(const void *key, const size_t keylen, const void *pt, const size_t tlen,
+			void *tag, size_t *taglen);
 const char *crypto_errstr(const unsigned long err);
 
 #define HASHSIZE 20
